Clean up Path planner stuff.
